Adds rust-like primitive type names
I see this naming scheme as more expressive and easy to understand especially to the newcomers. I know some gonna say that the size may differ in some platforms but let's be honest are realistic here, in 99% of the platform we target nothing changes in most of the types.
I've referred to INT64 as long long instead of long to ensure the 64bit-size in case of targeting ARM architecture.